+++ src/main/java/kitchenpos/application/TableService.java
@@ -4,8 +4,8 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.dao.order.OrderRepository;
 import kitchenpos.dao.order.OrderTableRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.order.OrderTable;
 import kitchenpos.dto.request.OrderTableRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
