+++ src/main/java/kitchenpos/dao/order/OrderLineItemDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.dao.order;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 
 import java.util.List;
 import java.util.Optional;
