+++ src/main/java/kitchenpos/dao/order/OrderLineItemRepository.java
@@ -1,7 +1,6 @@
 package kitchenpos.dao.order;
 
-import java.util.List;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
