+++ src/main/java/kitchenpos/dao/order/OrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.dao.order;
 
 import java.util.List;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
