+++ src/main/java/kitchenpos/dao/order/TableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao.order;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
