+++ src/main/java/kitchenpos/domain/order/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.Objects;
 import javax.persistence.Column;
