+++ src/main/java/kitchenpos/dto/request/OrderLineItemRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.request;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 
 public class OrderLineItemRequest {
