+++ src/main/java/kitchenpos/dto/response/OrderTableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.response;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.OrderTable;
 
 public class OrderTableResponse {
