+++ src/main/java/kitchenpos/dto/response/TableGroupResponse.java
@@ -2,9 +2,7 @@ package kitchenpos.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.TableGroup;
 
 public class TableGroupResponse {
